Add authorizations, issue fix

+++ training/views/batch.py
@@ -7,13 +7,15 @@ from django.shortcuts import get_object_or_404
 from django.urls import reverse
 from django.views.decorators.http import require_http_methods
 from django.views.generic import DetailView, FormView
+from django.utils.decorators import method_decorator
 
 from core import template_utils
-from core.utils import CustomDatatable
+from core.utils import CustomDatatable, validate_authorization
 from hubble.models import Batch, InternDetail, SubBatch
 from training.forms import BatchForm
 
 
+@method_decorator(validate_authorization(), name="dispatch")
 class BatchList(LoginRequiredMixin, FormView):
     """
     Timeline Template
@@ -23,6 +25,7 @@ class BatchList(LoginRequiredMixin, FormView):
     template_name = "batch/batch_list.html"
 
 
+@method_decorator(validate_authorization(), name="dispatch")
 class BatchDataTable(LoginRequiredMixin, CustomDatatable):
     """
     Batch Datatable
@@ -60,18 +63,18 @@ class BatchDataTable(LoginRequiredMixin, CustomDatatable):
         )
 
     def customize_row(self, row, obj):
-        buttons = (
-            template_utils.show_button(reverse("batch.detail", args=[obj.id]))
-            + template_utils.edit_button(reverse("batch.show", args=[obj.id]))
-            + template_utils.delete_button("deleteBatch('" + reverse("batch.delete", args=[obj.id]) + "')")
-        )
-        row[
-            "action"
-        ] = f'<div class="form-inline justify-content-center">{buttons}</div>'
+        buttons = (template_utils.show_button(reverse("batch.detail", args=[obj.id])))
+        if (self.request.user.is_admin_user):
+            buttons += (
+                template_utils.edit_button(reverse("batch.show", args=[obj.id]))
+                + template_utils.delete_button("deleteBatch('" + reverse("batch.delete", args=[obj.id]) + "')")
+            )
+        row["action"] = f'<div class="form-inline justify-content-center">{buttons}</div>'
         return
 
 
 @login_required()
+@validate_authorization()
 def create_batch(request):
     """
     Create Batch
@@ -96,6 +99,7 @@ def create_batch(request):
 
 
 @login_required()
+@validate_authorization()
 def batch_data(request, pk):
     """
     Batch Update Form Data
@@ -110,6 +114,7 @@ def batch_data(request, pk):
 
 
 @login_required()
+@validate_authorization()
 def update_batch(request, pk):
     """
     Update Batch
@@ -132,6 +137,7 @@ def update_batch(request, pk):
 
 
 @login_required()
+@validate_authorization()
 @require_http_methods(["DELETE"])  # This decorator ensures that the view function is only accessible through the DELETE HTTP method
 def delete_batch(request, pk):
     """
@@ -149,6 +155,7 @@ def delete_batch(request, pk):
         return JsonResponse({"message": "Error while deleting Batch!"}, status=500)
 
 
+@method_decorator(validate_authorization(), name="dispatch")
 class BatchDetails(LoginRequiredMixin, DetailView):
     model = Batch
     template_name = "sub_batch/sub_batch.html"
